auth-account/account-service/requirements.txt
fastapi == 0.95.0
uvicorn == 0.23.0
python-dotenv
sqlalchemy
pytz
tzlocal
supabase
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]